refactor: migrate book types to kiwi namespace and update related imports
