+++ apps/web/src/types/kiwi.ts
@@ -1,34 +1,41 @@
-import { BookData, Settings, ReadingRecord } from "./book";
+import { NavItem } from "@bookiwi/epubjs/types/navigation";
 
-export interface DiscussionProps {
-  id: string;
+export interface Settings {
+  isSinglePage: boolean;
+  fontFamily?: string;
+  fontSize?: number;
+  lineHeight?: number;
+  fontWeight?: number;
+}
+
+export interface BookmarkItem {
+  cfi: string;
+  timestamp: number;
+}
+
+export interface ReadingRecord {
+  currentCfi: string | null;
+  percentage: number | null;
+  bookmarks: BookmarkItem[];
+}
+
+export interface Metadata {
   title: string;
-  comments: number;
-  lastActive: string;
+  author: string;
+  publisher: string;
+  toc: NavItem[];
+  locations: string;
 }
 
-export interface EventProps {
-  id: string;
-  title: string;
-  date: string;
-  time: string;
-  participants: number;
+export interface BookData {
+  file: File;
+  coverImage: string | null;
+  metadata: Metadata;
 }
-
-export interface ActivityProps {
-  id: string;
-  user: string;
-  action: string;
-  content: string;
-  time: string;
-}
-
-export interface Member {
-  id: number;
-  name: string;
-  avatar: string;
-  progress: number;
-  lastActivityAt: string;
+export interface BookDataDB {
+  file: File;
+  coverImage: Blob | null;
+  metadata: Metadata;
 }
 
 export interface ParticipantType {
@@ -57,4 +64,31 @@ export interface Kiwi {
   participants: ParticipantType[];
   discussions?: DiscussionProps[];
   events?: EventProps[];
+}
+
+export interface KiwiDB extends Omit<Kiwi, "book"> {
+  book: BookDataDB;
+}
+
+export interface DiscussionProps {
+  id: string;
+  title: string;
+  comments: number;
+  lastActive: string;
+}
+
+export interface EventProps {
+  id: string;
+  title: string;
+  date: string;
+  time: string;
+  participants: number;
+}
+
+export interface ActivityProps {
+  id: string;
+  user: string;
+  action: string;
+  content: string;
+  time: string;
 }
